server/wan_api/requirements.txt
# Wan 2.1 I2V API - load model from Hugging Face
torch>=2.4.0
diffusers>=0.34.0
transformers>=4.45.0
accelerate>=0.33.0
flask>=3.0.0
pillow>=10.0.0
numpy>=1.24.0
